Move MDSpanViewTraits
